+++ src/store/note_record/input_note_record.rs
@@ -7,7 +7,7 @@ use miden_objects::{
     Digest,
 };
 
-use super::{NoteRecordDetails, NoteStatus};
+use super::{NoteRecordDetails, NoteStatus, OutputNoteRecord};
 use crate::errors::ClientError;
 
 // INPUT NOTE RECORD
@@ -174,16 +174,52 @@ impl TryInto<InputNote> for InputNoteRecord {
                 Ok(InputNote::new(note, proof.clone()))
             },
 
-            (None, _) => {
-                Err(ClientError::NoteError(miden_objects::NoteError::invalid_origin_index(
-                    "Input Note Record contains no inclusion proof".to_string(),
-                )))
+            (None, _) => Err(ClientError::NoteRecordError(
+                "Input Note Record contains no inclusion proof".to_string(),
+            )),
+            (_, None) => Err(ClientError::NoteRecordError(
+                "Input Note Record contains no metadata".to_string(),
+            )),
+        }
+    }
+}
+
+impl TryInto<Note> for InputNoteRecord {
+    type Error = ClientError;
+
+    fn try_into(self) -> Result<Note, Self::Error> {
+        match self.metadata {
+            Some(metadata) => {
+                let note_inputs = NoteInputs::new(self.details.inputs)?;
+                let note_recipient =
+                    NoteRecipient::new(self.details.serial_num, self.details.script, note_inputs);
+                let note = Note::new(self.assets, metadata, note_recipient);
+                Ok(note)
             },
-            (_, None) => {
-                Err(ClientError::NoteError(miden_objects::NoteError::invalid_origin_index(
-                    "Input Note Record contains no metadata".to_string(),
-                )))
-            },
+            None => Err(ClientError::NoteRecordError(
+                "Input Note Record contains no metadata".to_string(),
+            )),
+        }
+    }
+}
+
+impl TryFrom<OutputNoteRecord> for InputNoteRecord {
+    type Error = ClientError;
+
+    fn try_from(output_note: OutputNoteRecord) -> Result<Self, Self::Error> {
+        match output_note.details() {
+            Some(details) => Ok(InputNoteRecord {
+                assets: output_note.assets().clone(),
+                details: details.clone(),
+                id: output_note.id(),
+                inclusion_proof: output_note.inclusion_proof().cloned(),
+                metadata: Some(*output_note.metadata()),
+                recipient: output_note.recipient(),
+                status: output_note.status(),
+            }),
+            None => Err(ClientError::NoteError(miden_objects::NoteError::invalid_origin_index(
+                "Output Note Record contains no details".to_string(),
+            ))),
         }
     }
 }