+++ tests/integration/main.rs
@@ -12,7 +12,7 @@ use miden_client::{
         transactions::transaction_request::{
             PaymentTransactionData, TransactionRequest, TransactionTemplate,
         },
-        Client,
+        Client, NoteRelevance,
     },
     config::ClientConfig,
     errors::{ClientError, NodeRpcClientError},
@@ -964,6 +964,115 @@ async fn test_onchain_accounts() {
 
     assert_eq!(new_from_account_balance, from_account_balance - TRANSFER_AMOUNT);
     assert_eq!(new_to_account_balance, to_account_balance + TRANSFER_AMOUNT);
+}
+
+#[tokio::test]
+async fn test_get_consumable_notes() {
+    let mut client = create_test_client();
+
+    let (first_regular_account, second_regular_account, faucet_account_stub) =
+        setup(&mut client, AccountStorageMode::Local).await;
+
+    let from_account_id = first_regular_account.id();
+    let to_account_id = second_regular_account.id();
+    let faucet_account_id = faucet_account_stub.id();
+
+    //No consumable notes initially
+    assert!(client.get_consumable_notes(None).unwrap().is_empty());
+
+    // First Mint necesary token
+    let note = mint_note(&mut client, from_account_id, faucet_account_id, NoteType::OffChain).await;
+
+    // Check that note is consumable by the account that minted
+    assert!(!client.get_consumable_notes(None).unwrap().is_empty());
+    assert!(!client.get_consumable_notes(Some(from_account_id)).unwrap().is_empty());
+    assert!(client.get_consumable_notes(Some(to_account_id)).unwrap().is_empty());
+
+    consume_notes(&mut client, from_account_id, &[note]).await;
+
+    //After consuming there are no more consumable notes
+    assert!(client.get_consumable_notes(None).unwrap().is_empty());
+
+    // Do a transfer from first account to second account
+    let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
+    let tx_template = TransactionTemplate::PayToIdWithRecall(
+        PaymentTransactionData::new(Asset::Fungible(asset), from_account_id, to_account_id),
+        100,
+        NoteType::OffChain,
+    );
+    println!("Running P2IDR tx...");
+    let tx_request = client.build_transaction_request(tx_template).unwrap();
+    execute_tx_and_sync(&mut client, tx_request).await;
+
+    // Check that note is consumable by both accounts
+    let consumable_notes = client.get_consumable_notes(None).unwrap();
+    let relevant_accounts = &consumable_notes.first().unwrap().relevances;
+    assert_eq!(relevant_accounts.len(), 2);
+    assert!(!client.get_consumable_notes(Some(from_account_id)).unwrap().is_empty());
+    assert!(!client.get_consumable_notes(Some(to_account_id)).unwrap().is_empty());
+
+    // Check that the note is only consumable after block 100 for the account that sent the transaction
+    let from_account_relevance = relevant_accounts
+        .into_iter()
+        .find(|relevance| relevance.0 == from_account_id)
+        .unwrap()
+        .1;
+    assert_eq!(from_account_relevance, NoteRelevance::After(100));
+
+    // Check that the note is always consumable for the account that received the transaction
+    let to_account_relevance = relevant_accounts
+        .into_iter()
+        .find(|relevance| relevance.0 == to_account_id)
+        .unwrap()
+        .1;
+    assert_eq!(to_account_relevance, NoteRelevance::Always);
+}
+
+#[tokio::test]
+async fn test_get_output_notes() {
+    let mut client = create_test_client();
+
+    let (first_regular_account, _, faucet_account_stub) =
+        setup(&mut client, AccountStorageMode::Local).await;
+
+    let from_account_id = first_regular_account.id();
+    let faucet_account_id = faucet_account_stub.id();
+    let random_account_id = AccountId::from_hex("0x0123456789abcdef").unwrap();
+
+    //No output notes initially
+    assert!(client.get_output_notes(NoteFilter::All).unwrap().is_empty());
+
+    // First Mint necesary token
+    let note = mint_note(&mut client, from_account_id, faucet_account_id, NoteType::OffChain).await;
+
+    // Check that there was an output note but it wasn't consumed
+    assert!(client.get_output_notes(NoteFilter::Consumed).unwrap().is_empty());
+    assert!(!client.get_output_notes(NoteFilter::All).unwrap().is_empty());
+
+    consume_notes(&mut client, from_account_id, &[note]).await;
+
+    //After consuming, the note is returned when using the [NoteFilter::Consumed] filter
+    assert!(!client.get_output_notes(NoteFilter::Consumed).unwrap().is_empty());
+
+    // Do a transfer from first account to second account
+    let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
+    let tx_template = TransactionTemplate::PayToId(
+        PaymentTransactionData::new(Asset::Fungible(asset), from_account_id, random_account_id),
+        NoteType::OffChain,
+    );
+    println!("Running P2ID tx...");
+    let tx_request = client.build_transaction_request(tx_template).unwrap();
+
+    let output_note_id = tx_request.expected_output_notes()[0].id();
+
+    //Before executing, the output note is not found
+    assert!(client.get_output_note(output_note_id).is_err());
+
+    execute_tx_and_sync(&mut client, tx_request).await;
+
+    //After executing, the note is only found in output notes
+    assert!(client.get_output_note(output_note_id).is_ok());
+    assert!(client.get_input_note(output_note_id).is_err());
 }
 
 #[tokio::test]